refactor: move visibility component to its own directory
