+++ gravitee-apim-console-webui/src/management/api/documentation-v4/components/api-documentation-v4-visibility/api-documentation-v4-visibility.component.ts
